Fix subshell / subcommand usage in run-kibana.sh
`false || ( exit 1 )` doesn't actually exit the shell where `false`
failed, instead it creates a subshell and runs `exit 1` in it, so the
subshell exits with 1 but the main shell doesn't.

That works fine wiht `errexit` (the subshell exits with 1 and therefore
so does the main shell), but `run-kibana.sh` doesn't have that enabled,
so it doesn't.

Grouping the commands with `{` causes them to be executed in the main
shell and has the expected effect of exiting the shell.

---

Also, since `kibana` is the last thing that runs in this shell, it's not
useful to keep the shell around; we might as well `exec` kibana and make
it PID 1.