Kibana 4.4: Use elasticsearch.url for proxied reqs
Kibana 4.1 uses the credentials provided in the ES URL to authenticate
requests it proxies, which means users don't need to authenticate both
to Kibana and Elasticsearch when making a request to the latter through
the former (which isn't possible in the first place if both are using
basic auth with different credentials).

However, Kibana 4.4 does not do this by default, and proxies the headers
it received as-is to Elasticsearch, which doesn't work if Elasticsearch
uses different credentials (because the user's browser would need to
send the Kibana auth header to get through the Nginx instance proxying
Kibana *and* the Elasticsearch auth header to get through the one
proxying Elasticsearch).

So, this patch adds a patch for Kibana 4.4 to behave like Kibana 4.1 and
inject the credentials from `elasticsearch.url` when connecting to
Elasticsearch. It also adds some documentation to clarify what's going
on to users (even though nothing *new* is going on, at least as far as
Kibana 4.1 — which is presumably the only one in use on Aptible
currently — is concerned).